chore: add workflows and tidy tools
Restructure the tools code and create the workflows for release and
linting
