+++ .gitignore
@@ -1,2 +1,4 @@
 policy
 fanal
+
+trivy-plugin-mcp-*
