+++ Makefile
@@ -1,3 +1,27 @@
+PLATFORMS = linux/amd64 linux/arm64 darwin/amd64 darwin/arm64 windows/amd64
+OUTPUTS = $(patsubst %,%/trivy-plugin-mcp,$(PLATFORMS))
+
+.PHONY: clean
+clean:
+	rm -rf trivy-plugin-mcp*
+
+.PHONY: build
+build: clean $(OUTPUTS)
+%/trivy-plugin-mcp:
+	@echo "Building for $*..."
+	@mkdir -p $(dir $@); \
+	GOOS=$(word 1,$(subst /, ,$*)); \
+	GOARCH=$(word 2,$(subst /, ,$*)); \
+	CGO_ENABLED=0 GOOS=$$GOOS GOARCH=$$GOARCH go build -ldflags "-s -w" -o trivy-plugin-mcp-$$GOOS-$$GOARCH ./cmd/trivy-mcp/main.go; \
+	if [ $$GOOS = "windows" ]; then \
+		mv trivy-plugin-mcp-$$GOOS-$$GOARCH trivy-plugin-mcp-$$GOOS-$$GOARCH.exe; \
+		tar -czf trivy-plugin-mcp-$$GOOS-$$GOARCH.tar.gz plugin.yaml trivy-plugin-mcp-$$GOOS-$$GOARCH.exe LICENSE > /dev/null; \
+		rm trivy-plugin-mcp-$$GOOS-$$GOARCH.exe; \
+	else \
+		tar -czf trivy-plugin-mcp-$$GOOS-$$GOARCH.tar.gz plugin.yaml trivy-plugin-mcp-$$GOOS-$$GOARCH LICENSE > /dev/null; \
+		rm trivy-plugin-mcp-$$GOOS-$$GOARCH; \
+	fi
+
 .PHONY: add-plugin-manifest
 add-plugin-manifest:
 	@echo "Checking if plugin manifest exists..."
@@ -20,3 +44,16 @@ install-plugin: add-plugin-manifest
 run:
 	@echo "Running plugin code as 'trivy mcp -t sse -d'..."
 	@go run ./cmd/trivy-mcp -t sse -d
+
+.PHONY: lint
+lint:
+	@which golangci-lint || go install github.com/golangci/golangci-lint/cmd/golangci-lint@v1.64.2
+	@golangci-lint run --timeout 3m --verbose
+
+.PHONY: bump-manifest
+bump-manifest:
+	@[ $$NEW_VERSION ] || ( echo "env 'NEW_VERSION' is not set"; exit 1 )
+	@current_version=$$(cat plugin.yaml | grep 'version' | awk '{ print $$2}' | tr -d '"') ;\
+	echo Current version: $$current_version ;\
+	echo New version: $$NEW_VERSION ;\
+	$(SED) -i -e "s/$$current_version/$$NEW_VERSION/g" plugin.yaml ;\
